Fixing fails under Python 2.2
git-svn-id: svn://svn.twistedmatrix.com/svn/Twisted/trunk@487 bbbe8e31-12d6-0310-92fd-ac37d47ddeeb

+++ twisted/protocols/basic.py
@@ -77,7 +77,7 @@ class NetstringReceiver(protocol.Protocol):
                 raise NetstringParseError
         self.__data = self.__data[m.end():]
         if m.group(1):
-            self.length = self.length * (10**len(m.group(1))) + int(m.group(1))
+            self.length = self.length * (10**len(m.group(1))) + long(m.group(1))
         if m.group(2):
             self.__buffer = ''
             self.mode = DATA
